requirements.txt
PySDL2==0.9.8
PySDL2-dll==2.32.0